move some paths around
